src/mcp/requirements.txt
mcp[cli]
xarm-python-sdk
pymongo
numpy